requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
django==4.1
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   rest-pandas
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
lxml==4.9.1
    # via pandasdmx
numpy==1.23.1
    # via pandas
pandas==1.4.3
    # via
    #   pandasdmx
    #   rest-pandas
pandasdmx==1.9.0
    # via -r requirements.in
pydantic==1.9.1
    # via pandasdmx
python-dateutil==2.8.2
    # via pandas
pytz==2022.1
    # via
    #   djangorestframework
    #   pandas
requests==2.28.1
    # via pandasdmx
rest-pandas==1.1.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.11
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
